Change 'password (again)' for 'Confirm password
Change the silly 'password (again)' for 'Confirm password', because it describes
better its purpose, and this is the one of the used, standard message here.

New Subclass was established (AurRegistrationForm), because the original string
was established in the django-registration module, and it's no worth to modify it.
The password2 label of the mentionod class was refactored

The desited link should have established in the main urls.py according to the
accouns/register page and AurRegistrationForm subclass, so involving  it before
the original registration.views.register including.

Signed-off-by: Laszlo Papp <djszapi@archlinux.us>