Refactor AurRegistrationForm handling
This implementation reduces the dependencies, it means we could change only the
label, and not worry about anything else. It can happen when the
django-registration author changes the, or any other setting, inherit the
django-registration definition of password2, then change it on object
instanciation. This means the deletion of global variable 'attrs_dict' too.

Summary: The label is edited only, It's pretty safe unless the Forms api gets
altered significantly.

Signed-off-by: Laszlo Papp <djszapi@archlinux.us>