Change the activation handling during registration
This modification was needed because the activate/activation_complete handling
was changed, so it was needed to split the older activate.html for two separate
activation related templates.

activate.html is called at this momment when some error happened during
activation  either your activation link was incorrect, or the activation key for
your account has expired. So looking the view more account will never evaluate to
True in activate.html.

activation_complete.html is called when the activation was successful.

Summary: registration/activation.html is rendered if not valid,
registration/activation_complete.html is rendered if valid

Signed-off-by: Laszlo Papp <djszapi@archlinux.us>