Implement Remember me checkbox to the login page
This is a typical feature on websites, like in case of AUR too, so it was worth
to implement it. The Remember Me feature allows a user to select a Remember Me,
checkbox on the website login form.  When checked the user will be "remembered"
between browser sessions.

The implementation touched the new extended form class, the new view, extended
login template, and the url too

Signed-off-by: Laszlo Papp <djszapi@archlinux.us>