Implement OPENID facility on the site
Every new big site is using this feature, its the way things are going.
It also includes support for using a fixed OpenID server endpoint. Launchpad is
support at this commit only, but it will be extended later it's planned.

New dependency was established, named django-openid-auth, it can helps with
their implementation and documentation.

So this dependency was added to the pip-requirements.txt, and the related
modifications were established in the settings.py and in the urls.py

Signed-off-by: Laszlo Papp <djszapi@archlinux.us>