Redirect accounts/login/ while logged in
When a logged in user tried to load the accounts/login/ page, the real login
form was got meanwhile the actual user was logged in, so this was a small bug.

The solution was for it to redirect that page in this case to the 'aur-main'
one, which means the search page. So this redirections happens when the
user authentication is ok.

Signed-off-by: Laszlo Papp <djszapi@archlinux.us>