Add 'Registered: date' line into the Profile
It's good to know about a user when he or she registered and later it can be
shown to the other users who will be able to browse the profile of other users.

The implementation touched the aurprofile/profile.html template file, it was
needed to insert a new line into the profile form, with the builtind user
date_joined option with the real date in YYYY-MM-DD format.

Signed-off-by: Laszlo Papp <djszapi@archlinux.us>