Implement User Details page as /users/$username
It was a sense feature to implement, it can be seen on the current AUR too, when
you click for a username after you logged in.

The implementation touched the followings:

* The urls.py file, where object_details generic view was used with slug effect
* Establish the templates/aurprofile/user_detail.html related template file with
the users data, like lastname, firstname, username, registered date, email,
account type

Small remark: It's worth to redesign with css this page, because now it's very
pure.

Signed-off-by: Laszlo Papp <djszapi@archlinux.us>