Implement automatic new account adding to User group
The new users are automatically added to the default 'User' group after this
feature implementation. In case of more than 1000-1500 users it can be very
handy and comfortable :)

The implementation touched:

* aur/forms.py -> New AddGroupDefaultBackend subclass implementation as a
wrapper for the DefaultBackend of registration module, it was needed because the
register method should be overwritten with automatically adding the new accounts
into the defined 'User' group

* urls.py -> The DefaultBackend reference was changed for the newly created
AddGroupDefaultBackend url.

Signed-off-by: Laszlo Papp <djszapi@archlinux.us>