Add a warning message for trying to see user details without logging in
It's needed to implement as long it's implemented on the current AUR too.
So any trying to see the user details page as an anonymous user without logging
it, it won't work.

The implementation happened only in the desired template file, named
templated/aurprofile/user_detail.html with an if/else/endif is_authenticatiod
condition.

Signed-off-by: Laszlo Papp <djszapi@archlinux.us>