Implement /users/$username/packages url showing given user packages
This feature isn't implemented in the last format, it's just a very pure and
starting implementation for it that should be continued in the near future.

The packages from the desired user will be available for other users to see, and
check them.

The implementation is simple here, one url addition is needed for it, and to do
the template html for it, named templates/aur/package_list.html

Signed-off-by: Laszlo Papp <djszapi@archlinux.us>